+++ www/Library/admin_widgets/ImageEditor/res/jquery.ui.ImageEditor/jquery.ui.ImageEditor.js
@@ -1,0 +1,191 @@
+/**
+ * Виджет редактора изображения
+ * Query UI widget
+ * Copyright 2013 (C) Boolive
+ */
+(function($, document) {
+	$.widget("boolive.ImageEditor", $.boolive.AjaxWidget, {
+        // uri отображаемого объекта
+        _object: '',
+        _value: '',
+        _mouse_action: null, // Ожидаемое действие мышки
+        _curent_action: null, // Текущее действие
+        _resize_rect: null,
+
+        _create: function() {
+			$.boolive.AjaxWidget.prototype._create.call(this);
+            var self = this;
+            // uri объекта
+            self._object = this.element.attr('data-object');
+
+
+
+            self.element.on('mousedown'+this.eventNamespace, function(e){
+                if (self._mouse_action == 'resize'){
+                    self._make_selection(e);
+                }else{
+                    self._select();
+                }
+                self.element.attr('contentEditable', "false");
+                e.preventDefault();
+                e.stopPropagation();
+            }).on('mouseup'+this.eventNamespace, function(e){
+                e.preventDefault();
+                self.element.attr('contentEditable', "true");
+//                var sel = window.getSelection();
+//                sel.removeAllRanges();
+            }).on('click', function(e){
+                e.preventDefault();
+                e.stopPropagation();
+            }).on('mousemove', function(e){
+
+                if (self.element.hasClass('selected')){
+                    if (!self._curent_action){
+                        // Проверка попадания курсора в область ресайза
+                        var img_rect = self.element.offset();
+                        var rect = {
+                            top: img_rect.top + self.element.height() - 20,
+                            right: img_rect.left + self.element.width(),
+                            bottom: img_rect.top + self.element.height(),
+                            left: img_rect.left + self.element.width() - 20
+                        };
+                        if (self._isItersection(rect, {x: e.pageX, y: e.pageY})){
+                            // Доступен ресайз
+                            self.element.addClass('resizing');
+                            self._mouse_action = 'resize';
+                        }else{
+                            self.element.removeClass('resizing');
+                            self._mouse_action = false;
+                        }
+                    }
+                }else{
+                    self.element.removeClass('resizing');
+                    self._mouse_action = false;
+                }
+            });
+        },
+
+        _make_selection: function(e){
+            var self = this;
+            var img_rect = self.element.offset();
+            var width = self.element.width();
+            var height = self.element.height();
+            var pos = {x:e.pageX, y:e.pageY};
+            if (height==0) height = 0.01;
+            var prop = width/height;
+            var is_prop = true;
+            var namespace = '.resizer'+Math.random();
+            var dw, dh, w, h;
+            var resize_rect = $('<div> </div>').css({
+                position: 'absolute',
+                'z-index': 1000,
+                //border: '1px solid #000000',
+                'background-color': 'rgba(187, 219, 250, 0.5)',
+                top: img_rect.top,
+                width: width,
+                height: height,
+                left: img_rect.left,
+                cursor: 'se-resize'
+            });
+            $('body').append(resize_rect);
+            $(document).on('mousemove'+namespace, function(e){
+                dw = e.pageX - pos.x;
+                dh = e.pageY - pos.y;
+                if (is_prop){
+                    w = Math.max(10, width + Math.min(dh, dw));
+                    h = Math.round(w/prop);
+                }else{
+                    w = Math.max(10, width+dw);
+                    h = Math.max(10, height+dh);
+                }
+                resize_rect.css({
+                    width: w+'px',
+                    height: h+'px'
+                });
+            }).on('mouseup'+namespace, function(e){
+                $(this).off(namespace);
+                resize_rect.hide();
+                self.element.width(resize_rect.width());
+                self.element.height(resize_rect.height());
+
+                self._call('save_style', {
+                    save_style:{
+                        width: resize_rect.width()+'px',
+                        height: resize_rect.height()+'px'
+                    },
+                    object: self._object
+                });
+
+                self._curent_action = false;
+                self._select();
+            });
+        },
+
+        _isItersection: function(rect, point){
+             return rect.left<=point.x && rect.right>=point.x && rect.top<=point.y && rect.bottom>=point.y;
+        },
+
+        /**
+         * В режиме contenteditable своё событие клавитауры не получить,
+         * поэтому реагируем на событие родителя
+         * @param e Событие
+         * @param sel Выделение
+         */
+        after_keydown: function(e, sel){
+//            if (this._isInSelection(sel)){
+//                e.preventDefault();
+//                e.stopPropagation();
+//            }
+        },
+
+        after_keyup: function(e, selection){
+            if (this._isInSelection(selection)){
+                this._select();
+            }
+        },
+
+        /**
+         * Выделение (фокус) изображения
+         * @private
+         */
+        _select: function(){
+            this.element.attr('contentEditable', "true");
+            if (!this.element.hasClass('selected')){
+                this.before('setState', [{select:  this._object}]);
+                this.element.addClass('selected');
+
+                var sel = window.getSelection();
+                var range = document.createRange();
+                range.setStart(this.element.parent()[0], this.element.index());
+                range.collapse(true);
+                sel.removeAllRanges();
+                sel.addRange( range );
+            }
+        },
+
+        /**
+         * Отмена выделения (фокуса) изображения
+         * @param state
+         * @param changes
+         */
+        after_setState: function(state, changes){
+            if ('select' in changes && state.select!=this._object){
+                this.element.removeClass('selected');
+                this.element.removeClass('resizing');
+            }
+        },
+
+        /**
+         * Проверка, задевает ли текущее выделение элемент
+         * @param selection
+         * @return boolean
+         */
+        _isInSelection: function(selection){
+            if (selection.anchorNode.nodeType == 1 && selection.anchorOffset && selection.isCollapsed){
+                var node = selection.anchorNode.childNodes[selection.anchorOffset];
+                return this.element.is(node);
+            }
+            return false;
+        }
+	})
+})(jQuery, document);